Fix: Bug when using payment link.
Turn out that when changing the transaction reference code, there's a
field return_url that's used for non authenticated clients to see the
transaction result which is dependant of the reference code that was
passed at the point in time of this method.

This makes the controller method that displays the payment result, fail,
when comparing the token encoded in the return url, because the initial
token was calculated using the original reference code, and this one
with the modified.