feat: Added wait when client returns first.
Wompi sends the data from their servers, not in the client request back
to odoo, it only brings the internal id from wompi. This module has a
method that will ask to wompi manually for the information in the
transaction. And was fired away as soon as the client comes back and the
transaction hadn't been reported by wompi.

Now It waits 10 seconds before starting the process of updating the
transaction by calling Wompi api. This avoids an error of trying to
process the same transaction information twice.